backend/requirements.txt
fastapi
uvicorn
python-jose
passlib
pydantic
python-multipart
sqlalchemy
psycopg2-binary
pytest
httpx
